+++ changelog.md
@@ -18,6 +18,7 @@ Codion Change Log
 - EntityConditionModel renamed EntityTableConditionModel.
 - AttributeConditionModelFactory renamed EntityConditionModelFactory.
 - ForeignKeyConditionModel.ForeignKeyOperands removed.
+- EntityConditionModelFactory.conditionModel(ForeignKey) now returns ForeignKeyConditionModel. 
 ### is.codion.swing.framework.model
 - SwingForeignKeyConditionModel.equalComboBoxModel() and inSearchModel() moved to SwingForeignKeyOperands.
 ### is.codion.swing.framework.ui